restructured files, refined prompts
