+++ server.js
@@ -1,7 +1,9 @@
-const express = require('express');
-const cors = require('cors');
-const dotenv = require('dotenv');
-const ChatBotController = require('./controller/ChatBotController');
+// server.js
+
+const express = require("express");
+const cors = require("cors");
+const chatRoutes = require("./routes/ChatRoutes"); // Corrected path // This path is correct based on our refactoring
+const dotenv = require("dotenv");
 
 dotenv.config();
 
@@ -11,10 +13,7 @@ const port = process.env.PORT || 3000;
 app.use(cors());
 app.use(express.json());
 
-const chatBotController = new ChatBotController(process.env.GOOGLE_API_KEY);
-
-app.post('/api/ChatBot', (req, res) => chatBotController.handle(req, res));
-
+app.use("/api", chatRoutes);
 
 app.listen(port, () => {
   console.log(`Backend server running on http://localhost:${port}`);
